CSCMETAX-185: [REF] Re-org file tests
